Add Rack CORS Gem
Add in a gem called Rack CORS that will enable Cross-origin resource
sharing(CORS) calling because Rails in this application is only using
API and has not UI to it. Because of this in order for React-Flex to
access the the API either on the same server or on a different server it
will need to make a call to a that server outside of the application
normally this is block by the browser. However by enabling CORS the
application can allow other sites to calls its APIs that are not in the
same site.